requirements.txt
flask>=2.3.0,<4
requests>=2.28.0
python-dotenv>=1.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
